cleaned some unused code
